Refactor Database handlers for improved efficiency and maintainablility.
