+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "react-scripts": "^5.0.1"
+  }
+}
